requirement.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.1
django-paypal==2.1
gunicorn==23.0.0
idna==3.10
packaging==24.1
pillow==10.4.0
psycopg2==2.9.10
pytz==2024.2
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.8.0
